Enable Blackfire by default
Blackfire can be configured by defining the following Variables:

 * `env:BLACKFIRE_SERVER_ID`
 * `env:BLACKFIRE_SERVER_TOKEN`

It is safe to have Blackfire enabled even if it is not yet configured. Since
Blackfire is included with MECE it makes sense to have this enabled out of
the box, especially since it makes it a lot easier for the user to enable
and configure Blackfire.