Fix a case-sensitivity issue for non-case-sensitive environments
It looks as though at some point someone accidently committed a change to
SalesOrderView.xml on a system where case-sensitivity was a thing but did
so with the directory in the wrong directory. Or at the very least, it
was in a directory that wasn't correct anymore. The end result is that
although the file was more or less exactly the same (except for the
copyright dates) the following behavior would be observed:

On case-sensitive filesystems, things worked as expected. The tagging process
appears to favor the lower-case (incorrect, IMHO) diretory. This would
suggest to me that the tagging was done on a system that had
faux-case-sensitivity like a Mac?

On case-insensitive filesystems (or faux-case-sensitivity like a Mac) the result
is that the contents of both directories would end up in the same actual
directory on disk in those systems. Due to luck (or lack of luck?), since the
copyright changed in the lower-case directory, but the upper-case directory had
been created on disk first, the end result was immediately showing a change had
happened on disk. Running `git checkout -- <filename>` would not actually revert
the changes since git would not detect that any changes had actually happened.
People in these environments would be unable to create new branches, ever, as
this file was always showing as modified and thus git would refuse to change
branches.

The fix I chose for this was to move (using --force) the lower-case version of
the file onto the upper-case version of the file. The end result is that the
lower-case instance of the file is gone and the upper-case file will have the
updated copyright with the rest of the files in the upper-case directory.

This should ensure that case-insensitive or faux-case-sensitive filesystems
will handle things the same way as everyone else.

It also makes the repo more clean.